requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
pandas>=1.3.3
pyyaml>=5.4.1
tqdm>=4.62.2